+++ lib/bcdd/result/transitions/tracking/disabled.rb
@@ -6,22 +6,10 @@ module BCDD::Result::Transitions
       EnsureResult[yield]
     end
 
-    def self.reset!; end
-
     def self.record(result); end
 
     def self.record_and_then(_type, _data)
       yield
-    end
-
-    def self.reset_and_then!; end
-
-    class << self
-      private
-
-      def start(name, desc); end
-
-      def finish(result); end
     end
   end
 end