+++ test/bcdd/result/pattern_matching/deconstruct_test.rb
@@ -2,7 +2,7 @@
 
 require 'test_helper'
 
-class BCDD::PatternMatchingDeconstructKeysTest < Minitest::Test
+class BCDD::Result::PatternMatchingDeconstructTest < Minitest::Test
   Divide = ->(arg1, arg2) do
     arg1.is_a?(::Numeric) or return BCDD::Result::Failure(:invalid_arg, 'arg1 must be numeric')
     arg2.is_a?(::Numeric) or return BCDD::Result::Failure(:invalid_arg, 'arg2 must be numeric')
@@ -12,31 +12,28 @@ class BCDD::PatternMatchingDeconstructKeysTest < Minitest::Test
     BCDD::Result::Success(:division_completed, arg1 / arg2)
   end
 
-  test '#deconstruct_keys success' do
+  test '#deconstruct success' do
     result = Divide.call(10, 2)
 
-    assert_equal({ success: { division_completed: 5 } }, result.deconstruct_keys([]))
+    assert_equal [:division_completed, 5], result.deconstruct
 
     case result
-    in { failure: _ }
+    in BCDD::Result::Failure[_, _]
       raise
-    in { success: { division_completed: value } }
+    in BCDD::Result::Success[:division_completed, value]
       assert_equal 5, value
     end
   end
 
-  test '#deconstruct_keys failure' do
+  test '#deconstruct failure' do
     result = Divide.call(10, 0)
 
-    assert_equal(
-      { failure: { division_by_zero: 'arg2 must not be zero' } },
-      result.deconstruct_keys([])
-    )
+    assert_equal [:division_by_zero, 'arg2 must not be zero'], result.deconstruct
 
     case result
-    in { success: _ }
+    in BCDD::Result::Success[_, _]
       raise
-    in { failure: { division_by_zero: msg } }
+    in BCDD::Result::Failure[:division_by_zero, msg]
       assert_equal 'arg2 must not be zero', msg
     end
   end